requirements.txt
opencv-python
numpy
matplotlib
openvino-dev